requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
attrs==25.3.0
build==1.2.2.post1
certifi==2025.7.14
charset-normalizer==3.4.2
frozenlist==1.7.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
meraki==2.0.3
multidict==6.6.3
packaging==25.0
pluggy==1.6.0
propcache==0.3.2
Pygments==2.19.2
pyproject_hooks==1.2.0
pytest==8.4.1
python-dotenv==1.1.1
requests==2.32.4
rich==14.0.0
setuptools==78.1.1
urllib3==2.5.0
wheel==0.45.1
yarl==1.20.1
